Show camera name when offline
